Violation works

+++ E-Parking/src/main/java/com/eparking/entity/Violation.java
@@ -6,7 +6,7 @@ public class Violation {
 	private String plate;
 	private Date date;
 	private String description;
-	private double amount;
+	private int amount;
 	
 	public String getPlate() {
 		return plate;
@@ -26,10 +26,10 @@ public class Violation {
 	public void setDescription(String description) {
 		this.description = description;
 	}
-	public double getAmount() {
+	public int getAmount() {
 		return amount;
 	}
-	public void setAmount(double amount) {
+	public void setAmount(int amount) {
 		this.amount = amount;
 	}
